+++ packages/goat-fractal/src/twig/filters/render.js
@@ -1,9 +1,0 @@
-/**
- * Twig render filter, just return the input string
- * @export
- * @param {string} str
- * @returns {string}
- */
-module.exports = function(str) {
-  return str;
-};