+++ packages/goat-fractal/src/twig/filters/without.js
@@ -1,9 +1,0 @@
-/**
- * Twig without filter
- * @export
- * @param {string} str
- * @returns {string}
- */
-module.exports = function(str) {
-  return str;
-};