+++ packages/goat-modernizr/scripts/modernizr.js
@@ -1,4 +1,6 @@
-const _ = require('lodash');
+const map = require('lodash/map');
+const union = require('lodash/union');
+const isArray = require('lodash/isArray');
 const modernizr = require('customizr');
 
 module.exports = ({
@@ -6,7 +8,7 @@ module.exports = ({
   configuration,
 }) => {
   const { locations, js } = configuration;
-  let jsPaths = _.isArray(locations.javascript.src) ? _.map(locations.javascript.src, item => `${path}/${item}/**/*.js`) : locations.javascript.src;
+  const jsPaths = isArray(locations.javascript.src) ? map(locations.javascript.src, (item) => `${path}/${item}/**/*.js`) : locations.javascript.src;
   return modernizr({
     cache: true,
     devFile: false,
@@ -16,7 +18,7 @@ module.exports = ({
       'addTest',
       'html5printshiv',
       'testProp',
-      'fnBind'
+      'fnBind',
     ],
     uglify: js.minify || false,
     tests: js.modernizr.include || [],
@@ -24,11 +26,11 @@ module.exports = ({
     crawl: true,
     useBuffers: false,
     files: {
-      src: _.union([
-        locations.styles.src ? `${path}/${locations.styles.src}/**/*.s+(a|c)ss` : [], 
-        locations.javascript.libraries ? `!${path}/${locations.javascript.libraries}**/*.js` : [] 
-      ], jsPaths)
+      src: union([
+        locations.styles.src ? `${path}/${locations.styles.src}/**/*.s+(a|c)ss` : [],
+        locations.javascript.libraries ? `!${path}/${locations.javascript.libraries}**/*.js` : [],
+      ], jsPaths),
     },
-    customTests: []
+    customTests: [],
   });
-}
+};