+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## 1.9.
-- Fix: Updated dependencies
+## 1.10.0
+- Chore: Updated dependencies
+- Chore: removed `@geit/fractal`, due to poorly maintained dependencies and better alternatives
+- Feat: Moved Goat styles from node sass to dart sass
+- Feat: Removed `mkdirp` from goat babel in favor of fs.mkdir recursive
 
 ## 1.9.0
 ### Compile